feat: Rename circled icons to `IconCircle` so the same icons are grouped together
