+++ src/components/Icon/components.tsx
@@ -1,6 +1,6 @@
 export { ReactComponent as Account } from './assets/Account.svg';
 export { ReactComponent as Add } from './assets/Add.svg';
-export { ReactComponent as AddFilled } from './assets/AddFilled.svg';
+export { ReactComponent as AddCircle } from './assets/AddCircle.svg';
 export { ReactComponent as ArrowDown } from './assets/ArrowDown.svg';
 export { ReactComponent as ArrowLeft } from './assets/ArrowLeft.svg';
 export { ReactComponent as ArrowRight } from './assets/ArrowRight.svg';
@@ -14,9 +14,9 @@ export { ReactComponent as CaretDown } from './assets/CaretDown.svg';
 export { ReactComponent as CaretLeft } from './assets/CaretLeft.svg';
 export { ReactComponent as CaretRight } from './assets/CaretRight.svg';
 export { ReactComponent as CaretUp } from './assets/CaretUp.svg';
-export { ReactComponent as CircledCross } from './assets/CircledCross.svg';
-export { ReactComponent as CircledTick } from './assets/CircledTick.svg';
-export { ReactComponent as CircledWarning } from './assets/CircledWarning.svg';
+export { ReactComponent as CrossCircle } from './assets/CrossCircle.svg';
+export { ReactComponent as TickCircle } from './assets/TickCircle.svg';
+export { ReactComponent as WarningCircle } from './assets/WarningCircle.svg';
 export { ReactComponent as ComingSoon } from './assets/ComingSoon.svg';
 export { ReactComponent as Connect } from './assets/Connect.svg';
 export { ReactComponent as Copy } from './assets/Copy.svg';
