+++ src/components/Icon/urls.tsx
@@ -1,6 +1,6 @@
 export { default as Account } from './assets/Account.svg';
 export { default as Add } from './assets/Add.svg';
-export { default as AddFilled } from './assets/AddFilled.svg';
+export { default as AddCircle } from './assets/AddCircle.svg';
 export { default as ArrowDown } from './assets/ArrowDown.svg';
 export { default as ArrowLeft } from './assets/ArrowLeft.svg';
 export { default as ArrowRight } from './assets/ArrowRight.svg';
@@ -14,9 +14,9 @@ export { default as CaretDown } from './assets/CaretDown.svg';
 export { default as CaretLeft } from './assets/CaretLeft.svg';
 export { default as CaretRight } from './assets/CaretRight.svg';
 export { default as CaretUp } from './assets/CaretUp.svg';
-export { default as CircledCross } from './assets/CircledCross.svg';
-export { default as CircledTick } from './assets/CircledTick.svg';
-export { default as CircledWarning } from './assets/CircledWarning.svg';
+export { default as CrossCircle } from './assets/CrossCircle.svg';
+export { default as TickCircle } from './assets/TickCircle.svg';
+export { default as WarningCircle } from './assets/WarningCircle.svg';
 export { default as ComingSoon } from './assets/ComingSoon.svg';
 export { default as Connect } from './assets/Connect.svg';
 export { default as Copy } from './assets/Copy.svg';
